+++ src/DfaInference/SatinFolder.java
@@ -508,10 +508,10 @@ public class SatinFolder extends SatinObject implements SatinFolderInterface, Co
                     + bestDFA.getMDLComplexity()
                     + ", DFA complexity " + bestDFA.getDFAComplexity()
                     + " and " + bestDFA.nProductiveStates + " states");
-            System.out.println("Total time     = " + (double)(endTime - startTime) / 1000.0);
-            System.out.println("Initialization = " + (double)(initializationTime - startTime) / 1000.0);
-            System.out.println("Search         = " + (double)(searchTime - initializationTime) / 1000.0);
-            System.out.println("Solution       = " + (double)(endTime - searchTime) / 1000.0);
+            System.out.println("Total time     = " + (endTime - startTime) / 1000.0);
+            System.out.println("Initialization = " + (initializationTime - startTime) / 1000.0);
+            System.out.println("Search         = " + (searchTime - initializationTime) / 1000.0);
+            System.out.println("Solution       = " + (endTime - searchTime) / 1000.0);
 
             if (logger.isInfoEnabled()) {
                 logger.info("The winner is:\n" + bestDFA);
